Rename dir TabsList -> Tabs
Remove TabsList as default export, include Tabs export as module.
