+++ packages/svogg-components/src/lib/components/Tabs/index.js
@@ -2,5 +2,4 @@ import TabList from './TabList.svelte';
 import Tab from './Tab.svelte';
 import TabPanel from './TabPanel.svelte';
 
-export default TabList;
-export { Tab, TabPanel };
+export { TabList, Tab, TabPanel };
