+++ packages/svogg-components/src/lib/index.js
@@ -1,5 +1,5 @@
 export { default as ChipInput } from './components/ChipInput/index.js';
 export { default as Drawer } from './components/Drawer/index.js';
-export { default as TabList, Tab, TabPanel } from './components/TabList/index.js';
+export { TabList, Tab, TabPanel } from './components/Tabs/index.js';
 export { default as focusTrap } from './actions/focusTrap.js';
 export { default as clickOutside } from './actions/clickOutside.js';
